test: Reproduce one part of #19
@keorn here's one good reason for test driven development: Writing a
test will show you the amount of dependencies that you have to set up. A
lot of dependencies => bad. I say: "Tests will teach you how to
structure and modularize your code."

In this particular example it's the `router-link` component and using
the store in the top menu. Why do we need to stub the store for the
menu? :confused: