refactor: Use v-btn v-list-item `nuxt` property
See:
https://v2.vuetifyjs.com/en/components/buttons/#api
https://v2.vuetifyjs.com/en/components/lists/#api

@keorn: The vuetify docs are unnecessary complex with `v-for` loops on
a link collection. You are much better off writing down the template,
e.g. here you can see that we turn the `/standard` dropdown menu item
into a proper `<nuxt-link>`.

Regarding the appearance of the top menu on mobile devices: The best and
easiest solution is to use a sidebar on all devices.